requirements.txt
streamlit==1.38.0
joblib==1.4.2
keras==3.5.0
numpy==1.26.4
Pillow==10.4.0
librosa==0.10.2
praat-parselmouth==0.4.3
scipy==1.14.1
tensorflow==2.17.0
xgboost==3.0.0
python_speech_features
numpy
